add controller manager and event handler
